Moved the test over and got rid of the script.
