+++ api/opds_for_distributors.py
@@ -143,13 +143,23 @@ class OPDSForDistributorsAPI(BaseCirculationAPI):
             pass
 
     def checkout(self, patron, pin, licensepool, internal_format):
+        now = datetime.datetime.utcnow()
+
+        # This is not open-access content, but the loans are
+        # effectively of indefinite duration. To avoid confusing this
+        # with open-access content, grant a loan of 95 years duration.
+        #
+        # By this time, the US copyright on the work will have expired
+        # and it will be open-access content -- so something needs to
+        # change between now and then anyway.
+        the_far_future = now + datetime.timedelta(days=365*95)
         return LoanInfo(
             licensepool.collection,
             licensepool.data_source.name,
             licensepool.identifier.type,
             licensepool.identifier.identifier,
-            start_date=datetime.datetime.now(),
-            end_date=None,
+            start_date=now,
+            end_date=the_far_future,
         )
 
     def fulfill(self, patron, pin, licensepool, internal_format):
