+++ bin/database_reaper
@@ -1,0 +1,11 @@
+#!/usr/bin/env python
+"""Remove miscellaneous expired things (Credentials, CachedFeeds, etc.)
+from the database.
+"""
+import os
+import sys
+bin_dir = os.path.split(__file__)[0]
+package_dir = os.path.join(bin_dir, "..")
+sys.path.append(os.path.abspath(package_dir))
+from core.scripts import RunReaperMonitorsScript
+RunReaperMonitorsScript().run()
