--- bin/loan_reaper
@@ -1,9 +1,0 @@
-#!/usr/bin/env python
-"""Remove expired loans."""
-import os
-import sys
-bin_dir = os.path.split(__file__)[0]
-package_dir = os.path.join(bin_dir, "..")
-sys.path.append(os.path.abspath(package_dir))
-from scripts import LoanReaperScript
-LoanReaperScript().run()
